# apps/bff/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend-for-frontend API for the LLM Wiki project. Serves wiki sources and concepts with full-text search, import, and pipeline status.",
        "title": "LLM Wiki BFF API",
        "contact": {
            "name": "LLM Wiki Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/projects/{pid}/profile/guidance/{revision}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ProjectHeader": []
                    }
                ],
                "description": "Returns the exact validated guidance text and version metadata for a current bootstrap, candidate, or Active Profile reference.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get immutable Profile compile guidance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Immutable guidance artifact revision",
                        "name": "revision",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID; must match the URL",
                        "name": "X-Project-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.profileGuidanceArtifactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sources/{id}/annotation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ProjectHeader": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Get a source annotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AnnotationResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ProjectHeader": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Save or clear a source annotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Annotation",
                        "name": "annotation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AnnotationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AnnotationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "412": {
                        "description": "Precondition Failed",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AnnotationRequest": {
            "type": "object",
            "required": [
                "expected_generation"
            ],
            "properties": {
                "body": {
                    "type": "string"
                },
                "expected_generation": {
                    "type": "string"
                }
            }
        },
        "handler.AnnotationResponse": {
            "type": "object",
            "properties": {
                "ann_sha256": {
                    "type": "string"
                },
                "annotation_dirty": {
                    "type": "boolean"
                },
                "body": {
                    "type": "string"
                },
                "dirty": {
                    "type": "boolean"
                },
                "generation": {
                    "type": "string"
                },
                "has_annotation": {
                    "type": "boolean"
                },
                "lifecycle_status": {
                    "type": "string"
                },
                "raw_dirty": {
                    "type": "boolean"
                },
                "raw_path": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "v1.profileGuidanceArtifactData": {
            "type": "object",
            "properties": {
                "compile_guidance": {
                    "type": "string"
                },
                "input_digest": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                },
                "prompt_version": {
                    "type": "string"
                },
                "revision": {
                    "type": "string"
                },
                "schema_version": {
                    "type": "string"
                }
            }
        },
        "v1.profileGuidanceArtifactResponse": {
            "type": "object",
            "properties": {
                "guidance_artifact": {
                    "$ref": "#/definitions/v1.profileGuidanceArtifactData"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token issued by the standalone Auth service. Format: \\\"Bearer \u003ctoken\u003e\\\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ProjectHeader": {
            "description": "Project identifier header.",
            "type": "apiKey",
            "name": "X-Project-ID",
            "in": "header"
        }
    }
}